+++ shuup_tests/admin/test_picotable.py
@@ -47,8 +47,8 @@ def instance_id(instance):  # Test direct `display` callable
 def false_and_true():
     return [(False, "False"), (True, "True")]
 
-def get_pico(rf, model=None, columns=None):
-    get_default_shop()
+def get_pico(rf, admin_user, model=None, columns=None):
+    shop = get_default_shop()
     model = model or get_user_model()
     columns = columns or [
         Column("id", "Id", filter_config=Filter(), display=instance_id),
@@ -58,8 +58,10 @@ def get_pico(rf, model=None, columns=None):
         Column("is_active", "Is Active", filter_config=ChoicesFilter(choices=false_and_true)),  # `choices` callable
         Column("date_joined", "Date Joined", filter_config=DateRangeFilter())
     ]
-
-    request = apply_request_middleware(rf.get("/"))
+    admin_user.is_staff = True
+    admin_user.save()
+    shop.staff_members.add(admin_user)
+    request = apply_request_middleware(rf.get("/"), user=admin_user)
     return Picotable(
         request=request,
         columns=columns,
@@ -72,7 +74,7 @@ def get_pico(rf, model=None, columns=None):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_basic(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     data = pico.get_data({"perPage": 100, "page": 1})
     assert len(data["items"]) == get_user_model().objects.count()
 
@@ -80,7 +82,7 @@ def test_picotable_basic(rf, admin_user, regular_user):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_display(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     data = pico.get_data({"perPage": 100, "page": 1})
     for item in data["items"]:
         if item["id"] == admin_user.pk:
@@ -92,7 +94,7 @@ def test_picotable_display(rf, admin_user, regular_user):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_default_sort(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     data = pico.get_data({"perPage": 100, "page": 1})
     id = None
     for item in data["items"]:
@@ -104,7 +106,7 @@ def test_picotable_default_sort(rf, admin_user, regular_user):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_sort(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     data = pico.get_data({"perPage": 100, "page": 1, "sort": "-id"})
     id = None
     for item in data["items"]:
@@ -116,7 +118,7 @@ def test_picotable_sort(rf, admin_user, regular_user):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_invalid_sort(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     with pytest.raises(ValueError):
         data = pico.get_data({"perPage": 100, "page": 1, "sort": "-email"})
 
@@ -124,7 +126,7 @@ def test_picotable_invalid_sort(rf, admin_user, regular_user):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_choice_filter(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     data = pico.get_data({"perPage": 100, "page": 1, "filters": {"is_superuser": True}})
     assert len(data["items"]) == get_user_model().objects.filter(is_superuser=True).count()
 
@@ -132,7 +134,7 @@ def test_picotable_choice_filter(rf, admin_user, regular_user):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_text_filter(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     data = pico.get_data({"perPage": 100, "page": 1, "filters": {"email": admin_user.email}})
     assert len(data["items"]) == get_user_model().objects.filter(is_superuser=True).count()
 
@@ -140,15 +142,15 @@ def test_picotable_text_filter(rf, admin_user, regular_user):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_multi_filter(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     data = pico.get_data({"perPage": 100, "page": 1, "filters": {"username": "."}})
     assert len(data["items"]) == get_user_model().objects.count()
 
 
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
-def test_picotable_range_filter(rf, regular_user):
-    pico = get_pico(rf)
+def test_picotable_range_filter(rf, admin_user, regular_user):
+    pico = get_pico(rf, admin_user)
     one_day = datetime.timedelta(days=1)
     assert not empty_iterable(pico.get_data({"perPage": 100, "page": 1, "filters": {"date_joined": {"min": regular_user.date_joined - one_day}}})["items"])
     assert not empty_iterable(pico.get_data({"perPage": 100, "page": 1, "filters": {"date_joined": {"max": regular_user.date_joined + one_day}}})["items"])
@@ -161,7 +163,7 @@ def test_column_is_user_friendly():
 @pytest.mark.django_db
 @pytest.mark.usefixtures("regular_user")
 def test_picotable_no_mobile_link_for_missing_object_url(rf, admin_user, regular_user):
-    pico = get_pico(rf)
+    pico = get_pico(rf, admin_user)
     pico.get_object_url = lambda object: "http://www.fakeurl.com"
     data = pico.get_data({"perPage": 100, "page": 1})
     assert data["items"][0]["_linked_in_mobile"]
@@ -187,7 +189,7 @@ def test_choice_filter_with_default(rf, admin_user, regular_user):
 
     query = {"perPage": 100, "page": 1, "sort": "+id"}
 
-    pico_no_defaults = get_pico(rf, columns=(columns + is_active))
+    pico_no_defaults = get_pico(rf, admin_user, columns=(columns + is_active))
     data = pico_no_defaults.get_data(query)
 
     superuser_field = data["columns"][3]
@@ -200,7 +202,7 @@ def test_choice_filter_with_default(rf, admin_user, regular_user):
     user = get_user_model().objects.get(id=user_data["id"])
     assert user.is_active
 
-    pico_with_defaults = get_pico(rf, columns=(columns + is_active_with_default))
+    pico_with_defaults = get_pico(rf, admin_user, columns=(columns + is_active_with_default))
     data = pico_with_defaults.get_data(query)
     user_data = data["items"][0]
     user_with_defaults = get_user_model().objects.get(id=user_data["id"])
@@ -235,7 +237,7 @@ def test_picotable_correctly_sorts_translated_fields(rf, admin_user, regular_use
             filter_config=TextFilter(filter_field="translations__name")),
     ]
 
-    pico = get_pico(rf, model=Product, columns=columns)
+    pico = get_pico(rf, admin_user, model=Product, columns=columns)
 
     # Verify ascending sort
     sorted_products = pico.get_data({"perPage": 100, "page": 1, "sort": "+name"})
@@ -249,7 +251,7 @@ def test_picotable_correctly_sorts_translated_fields(rf, admin_user, regular_use
 
 
 @pytest.mark.django_db
-def test_mptt_filter(rf):
+def test_mptt_filter(rf, admin_user):
     parent_category = Category.objects.create(name="parent")
     child_category = Category.objects.create(name="child")
     parent_category.children.add(child_category)
@@ -262,7 +264,7 @@ def test_mptt_filter(rf):
             )
         )
     ]
-    pico = get_pico(rf, model=Category, columns=columns)
+    pico = get_pico(rf, admin_user, model=Category, columns=columns)
     data = pico.get_data({"perPage": 100, "page": 1, "filters": {"id": parent_category.id}})
     assert len(data["items"]) == 2