+++ doc/changelog.rst
@@ -14,6 +14,18 @@ Core
 
 - Add ``registration_shop`` for ``Contact``.
 - Add shop value to ``ContactGroup``.
+- Move Price display options to separate object called ``ContactGroupPriceDisplay``.
+
+Admin
+~~~~~
+
+- Move Contact Groups menu item under Contacts where it belongs.
+
+Tests
+~~~~~
+
+- Tests being run from admin now sets the shop properly to the session. If you
+  do not want this, add ``skip_session=True`` parameter for ``apply_request_middleware``.
 
 Shuup 1.6.8
 -----------